requirements.txt

pyspark>=3.3.0
azure-cosmos>=4.3.0
azure-storage-file-datalake>=12.6.0
python-dotenv>=0.19.0